aspeed: check for negative values returned by blk_getlength()
write_boot_rom() does not check for negative values. This is more a
problem for coverity than the actual code as the size of the flash
device is checked when the m25p80 object is created. So, we should not
even reach that code if there is anything wrong with the backing file.

Signed-off-by: Cédric Le Goater <clg@kaod.org>