net: add FTGMAC100 support
The FTGMAC100 device is an Ethernet controller configured and
controlled via a set of registers available on AHB bus. Packets are
transmitted and received using DMA ring buffers. There are a MAC and
MII/GMII interfaces.

The model is for the moment complete enough to satisfy the Linux
kernel but there are areas that need some more attention :

 - PHY registers
 - Automatic receive/transmit polling, no timers are used to poll the
   ring buffers descriptors. We just expect the qemu nic backend to
   push packets when they arrive.
 - CRC

And some are still on the TODO list (not used today in the drivers) :

 - Ethernet Address Filtering
 - Multicast
 - High priority queues
 - VLAN
 - NCSI, may be we could add n/w filters
 - etc.

The code is based on the Coldfire Fast Ethernet Controller model.

Signed-off-by: Cédric Le Goater <clg@kaod.org>
Tested-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Cédric Le Goater <clg@kaod.org>